Add files via upload
The first phase of the project is concerned with exploratory data analytics with two of these  implemented using the popular cutting-edge technology called the Python language. The projects are meant for the HR Team with clearly defined goals in terms of seeking to uncover the level of productivity, and behavioral trends such as satisfaction levels, ratings,   retention, and churn rate,  amongst others across all departments and strata of their workforce. It is implemented by name following a process-oriented and holistic approach to uncovering and generating hidden patterns across different variables available for our historical human resource dataset. The code snippets allow for the board of directors to get a bird's eye view assessment of the overall departments and hierarchy of employees, you will get to know which department is most productive, the level of employee satisfaction, retention, and churn rate is well-reviewed, promotional policies of the company as well as its level of fairness and career advancement programs, and some other socio-demographic factors such as age,  education, gender, marital status, education alongside location and geographical proximity are reviewed across all departments and strata of the company to facilitate for actionable-insights and reporting, on the part of the Board of Directors and Executives. In this project, you get to understand the importance of the operational impact of Data, for companies to stay relevant in the global marketplace.

The second phase comprises a handful sets of projects implemented using Microsoft Office 365, one of these ensuring a proper assessment of regional sales managers in an eCommerce historical database, using the  Microsoft Excel pivot tables, Slicers, and timelines to facilitate an actionable-decision-making process,  with user-friendly interface and experience guaranteed.